chore: move helper functions to utils
